@@ -1,9 +1,8 @@
-﻿namespace Building.A.RESTful.API.With.NancyFx
+﻿namespace Building.A.RESTful.API.With.NancyFx;
+
+public class Person
 {
-    public class Person
-    {
-        public int Id { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-    }
+    public int Id { get; set; }
+    public string FirstName { get; set; }
+    public string LastName { get; set; }
 }
