@@ -3,29 +3,26 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 
-namespace Building.A.RESTful.API.With.NancyFx
+namespace Building.A.RESTful.API.With.NancyFx;
+
+public class UserIdentity : ClaimsPrincipal
 {
-    public class UserIdentity : ClaimsPrincipal
+    internal const string HANDLEPERSON_PERMISSION = "HandlePerson";
+
+    public UserIdentity(string userName, Guid userIdentifier, IEnumerable<string> claims)
     {
-        internal const string HANDLEPERSON_PERMISSION = "HandlePerson";
-        public string UserName { get; private set; }
-        public Guid UserIdentifier { get; private set; }
+        UserName = userName;
+        UserIdentifier = userIdentifier;
+        var claimList = new List<Claim>();
+        foreach (string claim in claims) claimList.Add(new Claim(ClaimTypes.Role, claim));
+        AddIdentity(new ClaimsIdentity(claimList, "Basic"));
+    }
 
-        public UserIdentity(string userName, Guid userIdentifier, IEnumerable<string> claims)
-        {
-            UserName = userName;
-            UserIdentifier = userIdentifier;
-            List<Claim> claimList = new List<Claim>();
-            foreach (string claim in claims)
-            {
-                claimList.Add(new Claim(ClaimTypes.Role, claim));
-            }
-            AddIdentity(new ClaimsIdentity(claimList, "Basic"));
-        }
+    public string UserName { get; }
+    public Guid UserIdentifier { get; }
 
-        public bool HasClaim(string claim)
-        {
-            return Claims.Any(c => c.Value == claim);
-        }
+    public bool HasClaim(string claim)
+    {
+        return Claims.Any(c => c.Value == claim);
     }
 }
